+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
+### Added
+- valid `event.id` and `event.client` test cases
+
+### Fixed
+- `event.id` validations
 
 ## [2.0.0] - 2019-10-21
 ### Changed